requirements.txt
jira==3.0.1
rich==10.5.0
appdirs==1.4.4
pyyaml==5.4.1
